The fields related with money must allow decimals.

+++ application/migrations/2013_05_15_135148_create_employees_table.php
@@ -19,7 +19,7 @@ class Create_Employees_Table
 			$table->string('phone',11);
 			$table->string('address',200);
 			$table->string('bank_account',20);
-			$table->integer('salary');
+			$table->float('salary');
 			$table->boolean('active');
 			$table->date('startdate');
 			$table->date('stopdate');
