+++ application/migrations/2013_05_27_205711_create_payments_paysheet_table.php
@@ -14,7 +14,7 @@ class Create_Payments_Paysheet_Table {
 			$table->increments('id');
 			$table->integer('employee_id');
 			$table->integer('paysheet_id');
-			$table->integer('weekly_salary');
+			$table->float('weekly_salary');
 			$table->boolean('mo');
 			$table->boolean('tu');
 			$table->boolean('we');
@@ -22,17 +22,17 @@ class Create_Payments_Paysheet_Table {
 			$table->boolean('fr');
 			$table->boolean('sa');
 			$table->boolean('su');
-			$table->integer('feeding_bonus');
-			$table->integer('extra_hours');
-			$table->integer('production_bonus');
-			$table->integer('extra_raws');
-			$table->integer('others');
-			$table->integer('accrued_total');
-			$table->integer('sso');
-			$table->integer('faov');
-			$table->integer('forced_stop');
-			$table->integer('received_loans');
-			$table->integer('net_total');
+			$table->float('feeding_bonus');
+			$table->float('extra_hours');
+			$table->float('production_bonus');
+			$table->float('extra_raws');
+			$table->float('others');
+			$table->float('accrued_total');
+			$table->float('sso');
+			$table->float('faov');
+			$table->float('forced_stop');
+			$table->float('received_loans');
+			$table->float('net_total');
 		});
 	}
